add correct remove user

+++ src/models/user/user.controller.ts
@@ -8,10 +8,13 @@ import {
     ParseUUIDPipe,
     UseInterceptors,
     ClassSerializerInterceptor,
+    ForbiddenException,
 } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UserResponse } from '../../common/response/user.response';
 import { UserService } from './user.service';
+import { CurrentUser } from '../../common/decorators/current-user.decorator';
+import { JwtPayload } from '../../common/interfaces/jwt.interface';
 
 // todo: you use UserController in your application?
 @Controller('user')
@@ -40,7 +43,14 @@ export class UserController {
 
     @UseInterceptors(ClassSerializerInterceptor)
     @Delete(':id')
-    public async remove(@Param('id', ParseUUIDPipe) id: string): Promise<UserResponse> {
+    public async remove(
+        @Param('id', ParseUUIDPipe) id: string,
+        @CurrentUser() userJwtPayload: JwtPayload,
+    ): Promise<UserResponse> {
+        if (id !== userJwtPayload.id) {
+            throw new ForbiddenException();
+        }
+
         const user = await this.userService.remove(id);
 
         return new UserResponse(user);
